Move RepresentationResolver tests to separate file
